# tripplanner/src/main/java/com/example/tripplanner/application/dto/AiLogPageResponse.java
package com.example.tripplanner.application.dto;

import lombok.Builder;
import lombok.Data;
import java.util.List;

@Data
@Builder
public class AiLogPageResponse {
    private List<AiLogSummaryResponse> content;
    private int page;
    private int size;
    private long totalElements;
    private int totalPages;
    private boolean last;
}