# tripplanner/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: AI Trip Planner API
  description: >
    Production-ready REST API for the AI-powered Trip Planner system.
    Implements Clean Architecture with an AI Orchestration pipeline:
    User → AI Orchestrator → LLM → Validator → Database → Response.
  version: 1.0.0
  contact:
    name: Nhom1 Dev Team

servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
  - url: https://api.tripplanner.io/v1
    description: Production Server

tags:
  - name: Trips
    description: Core trip management and AI generation
  - name: Itineraries
    description: Day-by-day itinerary management
  - name: Activities
    description: Individual activity within an itinerary
  - name: AI Logs
    description: AI execution tracking and observability

paths:

  # ─────────────────────────────────────────────
  # TRIP ENDPOINTS
  # ─────────────────────────────────────────────

  /trips:
    post:
      tags:
        - Trips
      summary: Create a new trip
      operationId: createTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripRequest'
      responses:
        '201':
          description: Trip created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /trips/{tripId}:
    get:
      tags:
        - Trips
      summary: Get trip by ID
      operationId: getTripById
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Trip found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Trips
      summary: Update an existing trip
      operationId: updateTrip
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdateRequest'
      responses:
        '200':
          description: Trip updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /trips/{tripId}/generate:
    post:
      tags:
        - Trips
      summary: Generate an AI trip plan
      description: >
        Triggers the AI Orchestration pipeline:
        Validates input → Constructs prompt → Calls LLM → Validates response →
        Persists itinerary & activities → Returns full plan.
      operationId: generateAiTripPlan
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: AI trip plan generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/AiServiceError'

  /trips/{tripId}/regenerate:
    post:
      tags:
        - Trips
      summary: Regenerate (retry) the AI trip plan
      description: >
        Re-triggers the AI Orchestration pipeline for a trip that already has a plan.
        Increments retry count in AiLog. Previous itinerary is replaced.
      operationId: regenerateAiTripPlan
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateRequest'
      responses:
        '200':
          description: AI trip plan regenerated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/AiServiceError'

  # ─────────────────────────────────────────────
  # ITINERARY ENDPOINTS
  # ─────────────────────────────────────────────

  /trips/{tripId}/itineraries:
    get:
      tags:
        - Itineraries
      summary: Get all itineraries for a trip
      operationId: getItinerariesByTripId
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: List of itineraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItineraryResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /trips/{tripId}/itineraries/{itineraryId}:
    put:
      tags:
        - Itineraries
      summary: Update an itinerary
      operationId: updateItinerary
      parameters:
        - $ref: '#/components/parameters/TripId'
        - $ref: '#/components/parameters/ItineraryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryUpdateRequest'
      responses:
        '200':
          description: Itinerary updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────
  # ACTIVITY ENDPOINTS
  # ─────────────────────────────────────────────

  /itineraries/{itineraryId}/activities:
    post:
      tags:
        - Activities
      summary: Create a new activity in an itinerary
      operationId: createActivity
      parameters:
        - $ref: '#/components/parameters/ItineraryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /itineraries/{itineraryId}/activities/{activityId}:
    put:
      tags:
        - Activities
      summary: Update an activity
      operationId: updateActivity
      parameters:
        - $ref: '#/components/parameters/ItineraryId'
        - $ref: '#/components/parameters/ActivityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityUpdateRequest'
      responses:
        '200':
          description: Activity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Activities
      summary: Delete an activity
      operationId: deleteActivity
      parameters:
        - $ref: '#/components/parameters/ItineraryId'
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '204':
          description: Activity deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────
  # AI LOG ENDPOINTS
  # ─────────────────────────────────────────────

  /trips/{tripId}/ai-logs:
    get:
      tags:
        - AI Logs
      summary: Get all AI logs for a trip
      operationId: getAiLogsByTripId
      parameters:
        - $ref: '#/components/parameters/TripId'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AiLogStatus'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated list of AI logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiLogPageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /ai-logs/{logId}:
    get:
      tags:
        - AI Logs
      summary: Get AI log detail by ID
      operationId: getAiLogById
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI log detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiLogDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

# ─────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────

components:

  # ── Parameters ──────────────────────────────

  parameters:
    TripId:
      name: tripId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the trip

    ItineraryId:
      name: itineraryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the itinerary

    ActivityId:
      name: activityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the activity

  # ── Responses ───────────────────────────────

  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

    AiServiceError:
      description: Upstream AI service error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ─────────────────────────────────

  schemas:

    # ── Enums ──────────────────────────────

    TripStatus:
      type: string
      enum:
        - PLANNING
        - GENERATED
        - CONFIRMED
      description: Current lifecycle status of a trip

    AiLogStatus:
      type: string
      enum:
        - PENDING
        - SUCCESS
        - FAILED
        - RETRYING
      description: Execution status of an AI call

    ValidationType:
      type: string
      enum:
        - SCHEMA
        - CONTENT
        - HALLUCINATION_CHECK
        - NONE
      description: Type of validation applied to the AI response

    # ── User ───────────────────────────────

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Trip ───────────────────────────────

    TripRequest:
      type: object
      required:
        - userId
        - title
        - destination
        - startDate
        - endDate
      properties:
        userId:
          type: string
          format: uuid
          description: Owner user UUID
        title:
          type: string
          minLength: 3
          maxLength: 200
        destination:
          type: string
          minLength: 2
          maxLength: 300
        startDate:
          type: string
          format: date
          description: ISO 8601 date (YYYY-MM-DD)
        endDate:
          type: string
          format: date
          description: ISO 8601 date (YYYY-MM-DD); must be after startDate
        budget:
          type: number
          format: double
          minimum: 0

    TripUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        destination:
          type: string
          minLength: 2
          maxLength: 300
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
          format: double
          minimum: 0
        status:
          $ref: '#/components/schemas/TripStatus'

    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/TripStatus'
        createdAt:
          type: string
          format: date-time

    # ── AI Generation ──────────────────────

    GenerateRequest:
      type: object
      properties:
        preferences:
          type: string
          maxLength: 1000
          description: Optional free-text user preferences for the AI
        model:
          type: string
          description: Override the default LLM model (e.g., gpt-4o, gemini-1.5-pro)
          example: gpt-4o
        promptVersion:
          type: string
          description: Specific prompt template version to use
          example: v2.1

    RegenerateRequest:
      type: object
      properties:
        feedback:
          type: string
          maxLength: 1000
          description: Optional user feedback to guide the regeneration
        model:
          type: string
          description: Override the default LLM model for retry
        promptVersion:
          type: string
          description: Specific prompt template version to use for retry

    GenerateResponse:
      type: object
      properties:
        tripId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TripStatus'
        aiLogId:
          type: string
          format: uuid
          description: Reference to the AiLog entry for this generation call
        itineraries:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryResponse'
        generatedAt:
          type: string
          format: date-time

    # ── Itinerary ──────────────────────────

    ItineraryUpdateRequest:
      type: object
      properties:
        summary:
          type: string
          maxLength: 2000

    ItineraryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        dayNumber:
          type: integer
          minimum: 1
        date:
          type: string
          format: date
        summary:
          type: string
        activities:
          type: array
          items:
            $ref: '#/components/schemas/ActivityResponse'

    # ── Activity ───────────────────────────

    ActivityRequest:
      type: object
      required:
        - name
        - startTime
        - endTime
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        location:
          type: string
          maxLength: 500
        startTime:
          type: string
          format: time
          description: ISO 8601 local time (HH:mm:ss)
          example: "09:00:00"
        endTime:
          type: string
          format: time
          description: ISO 8601 local time (HH:mm:ss); must be after startTime
          example: "11:00:00"
        cost:
          type: number
          format: double
          minimum: 0
        activityOrder:
          type: integer
          minimum: 1
          description: Display order within the day

    ActivityUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        location:
          type: string
          maxLength: 500
        startTime:
          type: string
          format: time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          example: "11:00:00"
        cost:
          type: number
          format: double
          minimum: 0
        activityOrder:
          type: integer
          minimum: 1

    ActivityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        itineraryId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        location:
          type: string
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        cost:
          type: number
          format: double
        activityOrder:
          type: integer

    # ── AI Log ─────────────────────────────

    AiLogSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        model:
          type: string
        status:
          $ref: '#/components/schemas/AiLogStatus'
        totalTokens:
          type: integer
        executionTime:
          type: integer
          description: Execution time in milliseconds
        retryCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    AiLogDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        userInput:
          type: string
        prompt:
          type: string
        response:
          type: string
        model:
          type: string
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
        status:
          $ref: '#/components/schemas/AiLogStatus'
        retryCount:
          type: integer
        errorMessage:
          type: string
          nullable: true
        validationType:
          $ref: '#/components/schemas/ValidationType'
        executionTime:
          type: integer
          description: Execution time in milliseconds
        promptVersion:
          type: string
        createdAt:
          type: string
          format: date-time

    AiLogPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AiLogSummaryResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean

    # ── Error ──────────────────────────────

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          type: string
          description: Correlation ID for distributed tracing

    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          example: Unprocessable Entity
        message:
          type: string
          example: Validation failed
        path:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'

    FieldViolation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: string
          nullable: true